+++ doc/changelog.markdown
@@ -2,6 +2,8 @@ Lombok Changelog
 ----------------
 
 ### v0.11.1 (EDGE)
+* FEATURE: {Experimental} `@ExtensionMethod` is now available to add extensions to
+any type in the form of static methods that take as first parameter an object of that type. [Documentation on @ExtensionMethod](http://projectlombok.org/features/experimental/ExtensionMethod.html)
 * ENHANCEMENT: Small performance enhancements in `equals` and `hashCode`. [Issue #366](http://code.google.com/p/projectlombok/issues/detail?id=366)
 * BUGFIX: Eclipse refactor script 'rename method arguments' should work more often with lombok-affected methods.
 * BUGFIX: Using 'val' in an enhanced for loop did not work if the iterable was a raw type.